refactor(project): move files to feature-first directory structure
