+++ index.html
@@ -13,7 +13,7 @@
 <body>
 
 <script type=module src="src/main.js"></script>
-<script type=module src="src/renderConjugationTable.js"></script>
+<script type=module src="src/detail/renderConjugationTable.js"></script>
 <header class="banner">
     <img src="/images/lm-logo.svg" alt="Lexicon Meum logo" />
 </header>
