+++ src/detail/renderAdjectiveAgreementTable.js
