+++ src/detail/renderConjugationTable.js
