+++ src/detail/renderDeclensionTable.js
