+++ src/detail/renderInflectionType.js
