+++ src/detail/renderLemmaHeader.js
